requirements.txt
flutter
SpeechRecognition
sounddevice
numpy
openai
sumy
pymongo
python-dotenv
google.generativeai